requirements.txt
pdf2image>=1.16.0
Pillow>=8.0.0
pytesseract>=0.3.7
python-docx>=0.8.10
PyYAML>=5.4.1
tqdm>=4.50.0